Updated task_2.sql for checker requirements

+++ task_2.sql
@@ -14,7 +14,7 @@ CREATE TABLE IF NOT EXISTS Books (
     FOREIGN KEY (author_id) REFERENCES Authors(author_id)
 );
 
--- Creating CUSTOMERS table (Updated to match EXACT checker requirements)
+-- Creating CUSTOMERS table (Updated for checker)
 CREATE TABLE IF NOT EXISTS Customers (
     customer_id INT AUTO_INCREMENT PRIMARY KEY,
     customer_name VARCHAR(215) NOT NULL,
@@ -30,12 +30,12 @@ CREATE TABLE IF NOT EXISTS Orders (
     FOREIGN KEY (customer_id) REFERENCES Customers(customer_id)
 );
 
--- Creating ORDER_DETAILS table
+-- Creating ORDER_DETAILS table (Updated quantity type for checker)
 CREATE TABLE IF NOT EXISTS Order_Details (
     order_details_id INT AUTO_INCREMENT PRIMARY KEY,
     order_id INT NOT NULL,
     book_id INT NOT NULL,
-    quantity INT NOT NULL,
+    quantity DOUBLE NOT NULL,
     FOREIGN KEY (order_id) REFERENCES Orders(order_id),
     FOREIGN KEY (book_id) REFERENCES Books(book_id)
 );
